split lib and gradle plugin into separate sub-projects
