+++ settings.gradle.kts
@@ -1,2 +1,4 @@
 rootProject.name = "webidl-util"
 
+include("webidl-util")
+include("webidl-util-gradle")
