+++ webidl-util/src/main/kotlin/de/fabmax/webidl/generator/jni/java/JniFrameworkClassGenerator.kt
