+++ webidl-util/src/main/kotlin/de/fabmax/webidl/generator/jni/nat/JavaToNativeTypeMapping.kt
