+++ webidl-util/src/main/kotlin/de/fabmax/webidl/generator/ktjs/EmscriptenIdlGenerator.kt
