+++ webidl-util/src/main/kotlin/de/fabmax/webidl/model/IdlAttribute.kt
