+++ webidl-util/src/main/kotlin/de/fabmax/webidl/model/IdlDecorator.kt
