+++ webidl-util/src/main/kotlin/de/fabmax/webidl/model/IdlInterface.kt
