+++ webidl-util/src/main/kotlin/de/fabmax/webidl/parser/WebIdlParser.kt
