+++ webidl-util/src/test/resources/bad-enum.idl
